Refactor: Uncouple property map from Finder
