+++ corehq/motech/openmrs/finders.py
@@ -15,13 +15,11 @@ from operator import eq
 from pprint import pformat
 
 import six
-from jsonpath_rw import Child, parse, Fields, Slice, Where
 
 from corehq.motech.openmrs.finders_utils import (
     le_days_diff,
     le_levenshtein_percent,
 )
-from corehq.motech.openmrs.jsonpath import Cmp
 from corehq.motech.value_source import recurse_subclasses
 from dimagi.ext.couchdbkit import (
     BooleanProperty,
@@ -83,43 +81,6 @@ class PatientFinder(DocumentSchema):
         raise NotImplementedError
 
 
-# JsonpathValueMap is for comparing OpenMRS patients with CommCare
-# cases.
-#
-# The `jsonpath` attribute is used for retrieving values from an
-# OpenMRS patient and the `value_map` attribute is for converting
-# OpenMRS concept UUIDs to CommCare property values, if necessary.
-JsonpathValuemap = namedtuple('JsonpathValuemap', ['jsonpath', 'value_map'])
-
-
-def get_caseproperty_jsonpathvaluemap(jsonpath, value_source):
-    """
-    Used for updating _property_map to map case properties to OpenMRS
-    patient property-, attribute- and concept values.
-
-    i.e. Allows us to answer the question, "If we know the case property how
-    do we find the OpenMRS value?"
-
-    :param jsonpath: The path to a value in an OpenMRS patient JSON object
-    :param value_source: A case_config ValueSource instance
-    :return: A single-item dictionary with the name of the case
-             property as key, and a JsonpathValuemap as value. If
-             value_source is a constant, then there is no corresponding
-             case property, so the function returns an empty dictionary
-    """
-    if value_source['doc_type'] == 'ConstantString':
-        return {}
-    if value_source['doc_type'] == 'CaseProperty':
-        return {value_source['case_property']: JsonpathValuemap(jsonpath, {})}
-    if value_source['doc_type'] == 'CasePropertyMap':
-        value_map = {v: k for k, v in value_source['value_map'].items()}
-        return {value_source['case_property']: JsonpathValuemap(jsonpath, value_map)}
-    raise ValueError(
-        '"{}" is not a recognised ValueSource for setting OpenMRS patient values from CommCare case properties. '
-        'Please check your OpenMRS case config.'.format(value_source['doc_type'])
-    )
-
-
 PatientScore = namedtuple('PatientScore', ['patient', 'score'])
 
 
@@ -179,113 +140,6 @@ class WeightedPropertyPatientFinder(PatientFinder):
         super(WeightedPropertyPatientFinder, self).__init__(*args, **kwargs)
         self._property_map = {}
 
-    def set_property_map(self, case_config):
-        """
-        Set self._property_map to map OpenMRS properties and attributes
-        to case properties.
-        """
-        # Example value of case_config::
-        #
-        #     {
-        #         "person_properties": {
-        #             "birthdate": {
-        #                 "doc_type": "CaseProperty",
-        #                 "case_property": "dob"
-        #             }
-        #         },
-        #         "person_preferred_name": {
-        #             "givenName": {
-        #                 "doc_type": "CaseProperty",
-        #                 "case_property": "given_name"
-        #             },
-        #             "familyName": {
-        #                 "doc_type": "CaseProperty",
-        #                 "case_property": "family_name"
-        #             }
-        #         },
-        #         "person_preferred_address": {
-        #             "address1": {
-        #                 "doc_type": "CaseProperty",
-        #                 "case_property": "address_1"
-        #             },
-        #             "address2": {
-        #                 "doc_type": "CaseProperty",
-        #                 "case_property": "address_2"
-        #             }
-        #         },
-        #         "person_attributes": {
-        #             "c1f4239f-3f10-11e4-adec-0800271c1b75": {
-        #                 "doc_type": "CaseProperty",
-        #                 "case_property": "caste"
-        #             },
-        #             "c1f455e7-3f10-11e4-adec-0800271c1b75": {
-        #                 "doc_type": "CasePropertyMap",
-        #                 "case_property": "class",
-        #                 "value_map": {
-        #                     "sc": "c1fcd1c6-3f10-11e4-adec-0800271c1b75",
-        #                     "general": "c1fc20ab-3f10-11e4-adec-0800271c1b75",
-        #                     "obc": "c1fb51cc-3f10-11e4-adec-0800271c1b75",
-        #                     "other_caste": "c207073d-3f10-11e4-adec-0800271c1b75",
-        #                     "st": "c20478b6-3f10-11e4-adec-0800271c1b75"
-        #                 }
-        #             }
-        #         }
-        #         // ...
-        #     }
-        #
-        for person_prop, value_source in case_config['person_properties'].items():
-            jsonpath = parse('person.' + person_prop)
-            self._property_map.update(get_caseproperty_jsonpathvaluemap(jsonpath, value_source))
-
-        for attr_uuid, value_source in case_config['person_attributes'].items():
-            # jsonpath_rw offers programmatic JSONPath expressions. For details on how to create JSONPath
-            # expressions programmatically see the
-            # `jsonpath_rw documentation <https://github.com/kennknowles/python-jsonpath-rw#programmatic-jsonpath>`__
-            #
-            # The `Where` JSONPath expression "*jsonpath1* `where` *jsonpath2*" returns nodes matching *jsonpath1*
-            # where a child matches *jsonpath2*. `Cmp` does a comparison in *jsonpath2*. It accepts a
-            # comparison operator and a value. The JSONPath expression below is the equivalent of::
-            #
-            #     (person.attributes[*] where attributeType.uuid eq attr_uuid).value
-            #
-            # This `for` loop will let us extract the person attribute values where their attribute type UUIDs
-            # match those configured in case_config['person_attributes']
-            jsonpath = Child(
-                Where(
-                    Child(Child(Fields('person'), Fields('attributes')), Slice()),
-                    Cmp(Child(Fields('attributeType'), Fields('uuid')), eq, attr_uuid)
-                ),
-                Fields('value')
-            )
-            self._property_map.update(get_caseproperty_jsonpathvaluemap(jsonpath, value_source))
-
-        for name_prop, value_source in case_config['person_preferred_name'].items():
-            jsonpath = parse('person.preferredName.' + name_prop)
-            self._property_map.update(get_caseproperty_jsonpathvaluemap(jsonpath, value_source))
-
-        for addr_prop, value_source in case_config['person_preferred_address'].items():
-            jsonpath = parse('person.preferredAddress.' + addr_prop)
-            self._property_map.update(get_caseproperty_jsonpathvaluemap(jsonpath, value_source))
-
-        for id_type_uuid, value_source in case_config['patient_identifiers'].items():
-            if id_type_uuid == 'uuid':
-                jsonpath = parse('uuid')
-            else:
-                # The JSONPath expression below is the equivalent of::
-                #
-                #     (identifiers[*] where identifierType.uuid eq id_type_uuid).identifier
-                #
-                # Similar to `person_attributes` above, this will extract the person identifier values where
-                # their identifier type UUIDs match those configured in case_config['patient_identifiers']
-                jsonpath = Child(
-                    Where(
-                        Child(Fields('identifiers'), Slice()),
-                        Cmp(Child(Fields('identifierType'), Fields('uuid')), eq, id_type_uuid)
-                    ),
-                    Fields('identifier')
-                )
-            self._property_map.update(get_caseproperty_jsonpathvaluemap(jsonpath, value_source))
-
     def get_score(self, patient, case):
         """
         Return the sum of weighted properties to give an OpenMRS
@@ -315,9 +169,10 @@ class WeightedPropertyPatientFinder(PatientFinder):
         with a confidence score >= self.threshold
         """
         from corehq.motech.openmrs.logger import logger
+        from corehq.motech.openmrs.openmrs_config import get_property_map
         from corehq.motech.openmrs.repeater_helpers import search_patients
 
-        self.set_property_map(case_config)
+        self._property_map = get_property_map(case_config)
 
         candidates = {}  # key on OpenMRS UUID to filter duplicates
         for prop in self.searchable_properties:
